use Custom Hook to implement GET response
