+++ src/User.js
@@ -1,54 +1,16 @@
 import React, { useState, useEffect, useReducer } from "react";
 import axios from "axios";
+import useAsync from "./useAsync";
 
-function reducer(state, { type, data, error }) {
-  switch (type) {
-    case "LOADING":
-      return {
-        loading: true,
-        data: null,
-        error: null,
-      };
-    case "SUCCESS":
-      return {
-        loading: false,
-        data: data,
-        error: null,
-      };
-    case "ERROR":
-      return {
-        loading: false,
-        data: null,
-        error: error,
-      };
-    default:
-      throw new Error(`Unhandled action type: ${action.type}`);
-  }
+async function getUsers() {
+  const response = await axios.get(
+    "https://jsonplaceholder.typicode.com/users"
+  );
+  return response.data;
 }
 
 function Users() {
-  const [state, dispatch] = useReducer(reducer, {
-    loading: false,
-    data: null,
-    error: null,
-  });
-
-  const fetchUsers = async () => {
-    dispatch({ type: "LOADING" });
-    try {
-      const response = await axios.get(
-        "https://jsonplaceholder.typicode.com/users"
-      );
-      dispatch({ type: "SUCCESS", data: response.data });
-    } catch (e) {
-      dispatch({ type: "ERROR", error: e });
-    }
-  };
-
-  useEffect(() => {
-    fetchUsers();
-  }, []);
-
+  const [state, refetch] = useAsync(getUsers, []);
   const { loading, data: users, error } = state;
 
   if (loading) return <div>로딩중</div>;
@@ -64,7 +26,7 @@ function Users() {
           </li>
         ))}
       </ul>
-      <button onClick={fetchUsers}>다시 불러오기</button>
+      <button onClick={refetch}>다시 불러오기</button>
     </>
   );
 }
